refactor: Change likeCount and commentCount from int to long

+++ src/main/java/synapps/resona/api/socialMedia/dto/feed/FeedDto.java
@@ -22,8 +22,8 @@ public class FeedDto {
   private final String content;
   private final List<FeedMediaDto> images;
 
-  private final int likeCount;
-  private final int commentCount;
+  private final long likeCount;
+  private final long commentCount;
 
   private boolean hasLiked;
   private boolean hasScraped;
